Part of https://github.com/JuliaAlgebra/MultivariatePolynomials.jl/issues/194

## Benchmark 1

|                    |       Time |      Alloc |     Memory |
|--------------------|------------|------------|------------|
|    SIMDPolynomials |   3.052 ms |      86958 |   2.59 MiB |
| DynamicPolynomials |   4.946 ms |     107489 |   3.84 MiB |
|   TypedPolynomials |   3.139 ms |      86455 |   2.58 MiB |

## Benchmark 2

|                    |       Time |      Alloc |     Memory |
|--------------------|------------|------------|------------|
|    SIMDPolynomials |  21.231 μs |        623 |  25.92 KiB |
| DynamicPolynomials | 151.213 μs |       2656 | 161.92 KiB |
|   TypedPolynomials |  47.272 μs |        823 |  35.17 KiB |

## Benchmark 3

|                    |       Time |      Alloc |     Memory |
|--------------------|------------|------------|------------|
|    SIMDPolynomials |   2.715 ms |      76147 |   3.05 MiB |
| DynamicPolynomials |  13.290 ms |     222506 |  13.10 MiB |
|   TypedPolynomials |   2.509 ms |      66448 |   3.01 MiB |

Recommended by @YingboMa